+++ bukkit/src/main/java/de/maxhenkel/voicechat/config/ServerConfig.java
@@ -20,9 +20,11 @@ public class ServerConfig {
     public final ConfigEntry<Boolean> allowRecording;
     public final ConfigEntry<Boolean> spectatorInteraction;
     public ConfigEntry<Boolean> spectatorPlayerPossession;
+    public ConfigEntry<Boolean> forceVoiceChat;
+    public ConfigEntry<Integer> loginTimeout;
 
     public ServerConfig(ConfigBuilder builder) {
-        voiceChatPort = builder.integerEntry("port", 24454, 0, 65535);
+        voiceChatPort = builder.integerEntry("port", 24454, -1, 65535);
         voiceChatBindAddress = builder.stringEntry("bind_address", "");
         voiceChatDistance = builder.doubleEntry("max_voice_distance", 48D, 1D, 1_000_000D);
         voiceChatFadeDistance = builder.doubleEntry("min_voice_distance", 4D, 1D, 1_000_000D);
@@ -37,6 +39,8 @@ public class ServerConfig {
         allowRecording = builder.booleanEntry("allow_recording", true);
         spectatorInteraction = builder.booleanEntry("spectator_interaction", false);
         spectatorPlayerPossession = builder.booleanEntry("spectator_player_possession", false);
+        forceVoiceChat = builder.booleanEntry("force_voice_chat", false);
+        loginTimeout = builder.integerEntry("login_timeout", 10_000, 100, Integer.MAX_VALUE);
     }
 
     public enum Codec {