+++ bukkit/src/main/java/de/maxhenkel/voicechat/voice/server/ServerVoiceEvents.java
@@ -1,16 +1,18 @@
 package de.maxhenkel.voicechat.voice.server;
 
 import de.maxhenkel.voicechat.Voicechat;
-import de.maxhenkel.voicechat.command.VoiceChatCommands;
 import de.maxhenkel.voicechat.net.NetManager;
 import de.maxhenkel.voicechat.net.RequestSecretPacket;
 import de.maxhenkel.voicechat.net.SecretPacket;
+import de.maxhenkel.voicechat.permission.PermissionManager;
 import de.maxhenkel.voicechat.plugins.PluginManager;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.TextDecoration;
+import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
+import org.bukkit.event.player.PlayerLoginEvent;
 import org.bukkit.event.player.PlayerQuitEvent;
 
 import java.util.Map;
@@ -29,7 +31,7 @@ public class ServerVoiceEvents implements Listener {
             Voicechat.LOGGER.warn("Running in offline mode - Voice chat encryption is not secure!");
         }
 
-        server = new Server(Voicechat.SERVER_CONFIG.voiceChatPort.get(), mcServer);
+        server = new Server(mcServer);
         server.start();
         PluginManager.instance().onServerStarted(mcServer);
     }
@@ -77,17 +79,35 @@ public class ServerVoiceEvents implements Listener {
             return;
         }
         server.getPlayerStateManager().onPlayerCompatibilityCheckSucceeded(player);
-        if (!player.hasPermission(VoiceChatCommands.CONNECT_PERMISSION)) {
+
+        if (!player.hasPermission(PermissionManager.CONNECT_PERMISSION)) {
             Voicechat.LOGGER.info("Player {} has no permission to connect to the voice chat", player.getName());
             return;
         }
 
         UUID secret = server.getSecret(player.getUniqueId());
-
-        boolean hasGroupPermission = player.hasPermission(VoiceChatCommands.GROUPS_PERMISSION);
-
-        NetManager.sendToClient(player, new SecretPacket(player, secret, hasGroupPermission, Voicechat.SERVER_CONFIG));
+        NetManager.sendToClient(player, new SecretPacket(player, secret, Voicechat.SERVER_CONFIG));
         Voicechat.LOGGER.info("Sent secret to {}", player.getName());
+    }
+
+    @EventHandler
+    public void playerLoggedIn(PlayerLoginEvent event) {
+        if (!Voicechat.SERVER_CONFIG.forceVoiceChat.get()) {
+            return;
+        }
+        Player player = event.getPlayer();
+
+        Bukkit.getScheduler().runTaskLater(Voicechat.INSTANCE, () -> {
+            if (!player.isOnline()) {
+                return;
+            }
+            if (!isCompatible(player)) {
+                player.kickPlayer("You need %s %s to play on this server".formatted(
+                        Voicechat.INSTANCE.getDescription().getName(),
+                        Voicechat.INSTANCE.getDescription().getVersion()
+                ));
+            }
+        }, Voicechat.SERVER_CONFIG.loginTimeout.get() / 50L);
     }
 
     @EventHandler
@@ -98,7 +118,7 @@ public class ServerVoiceEvents implements Listener {
         }
 
         server.disconnectClient(event.getPlayer().getUniqueId());
-        Voicechat.LOGGER.info("Disconnecting client " + event.getPlayer().getName());
+        Voicechat.LOGGER.info("Disconnecting client {}", event.getPlayer().getName());
     }
 
     public Server getServer() {