+++ common/src/main/java/de/maxhenkel/voicechat/voice/server/Server.java
@@ -5,8 +5,10 @@ import de.maxhenkel.voicechat.api.RawUdpPacket;
 import de.maxhenkel.voicechat.api.VoicechatSocket;
 import de.maxhenkel.voicechat.api.events.SoundPacketEvent;
 import de.maxhenkel.voicechat.debug.CooldownTimer;
+import de.maxhenkel.voicechat.permission.PermissionManager;
 import de.maxhenkel.voicechat.plugins.PluginManager;
 import de.maxhenkel.voicechat.voice.common.*;
+import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.dedicated.DedicatedServer;
 import net.minecraft.server.level.ServerPlayer;
@@ -36,9 +38,15 @@ public class Server extends Thread {
 
     public Server(MinecraftServer server) {
         if (server instanceof DedicatedServer) {
-            this.port = Voicechat.SERVER_CONFIG.voiceChatPort.get();
+            int configPort = Voicechat.SERVER_CONFIG.voiceChatPort.get();
+            if (configPort < 0) {
+                Voicechat.LOGGER.info("Using the Minecraft servers port as voice chat port");
+                port = server.getPort();
+            } else {
+                port = configPort;
+            }
         } else {
-            this.port = 0;
+            port = 0;
         }
         this.server = server;
         socket = PluginManager.instance().getSocketImplementation(server);
@@ -173,6 +181,12 @@ public class Server extends Thread {
                     if (message.getPacket() instanceof MicPacket packet) {
                         ServerPlayer player = server.getPlayerList().getPlayer(playerUUID);
                         if (player == null) {
+                            continue;
+                        }
+                        if (!PermissionManager.INSTANCE.SPEAK_PERMISSION.hasPermission(player)) {
+                            CooldownTimer.run("muted-" + playerUUID, () -> {
+                                player.displayClientMessage(new TranslatableComponent("message.voicechat.no_speak_permission"), true);
+                            });
                             continue;
                         }
                         PlayerState state = playerStateManager.getState(player.getUUID());