+++ common/src/main/java/de/maxhenkel/voicechat/voice/server/ServerVoiceEvents.java
@@ -5,6 +5,7 @@ import de.maxhenkel.voicechat.VoicechatClient;
 import de.maxhenkel.voicechat.intercompatibility.CommonCompatibilityManager;
 import de.maxhenkel.voicechat.net.NetManager;
 import de.maxhenkel.voicechat.net.SecretPacket;
+import de.maxhenkel.voicechat.permission.PermissionManager;
 import de.maxhenkel.voicechat.plugins.PluginManager;
 import net.minecraft.ChatFormatting;
 import net.minecraft.Util;
@@ -17,6 +18,8 @@ import net.minecraft.server.level.ServerPlayer;
 
 import javax.annotation.Nullable;
 import java.util.Map;
+import java.util.Timer;
+import java.util.TimerTask;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -29,6 +32,7 @@ public class ServerVoiceEvents {
         clientCompatibilities = new ConcurrentHashMap<>();
         PluginManager.instance().init();
         CommonCompatibilityManager.INSTANCE.onServerStarting(this::serverStarting);
+        CommonCompatibilityManager.INSTANCE.onPlayerLoggedIn(this::playerLoggedIn);
         CommonCompatibilityManager.INSTANCE.onPlayerLoggedOut(this::playerLoggedOut);
         CommonCompatibilityManager.INSTANCE.onServerStopping(this::serverStopping);
 
@@ -93,9 +97,46 @@ public class ServerVoiceEvents {
             return;
         }
         server.getPlayerStateManager().onPlayerCompatibilityCheckSucceeded(player);
+
+        if (!PermissionManager.INSTANCE.CONNECT_PERMISSION.hasPermission(player)) {
+            Voicechat.LOGGER.info("Player {} has no permission to connect to the voice chat", player.getDisplayName().getString());
+            return;
+        }
+
         UUID secret = server.getSecret(player.getUUID());
         NetManager.sendToClient(player, new SecretPacket(player, secret, server.getPort(), Voicechat.SERVER_CONFIG));
-        Voicechat.LOGGER.info("Sent secret to " + player.getDisplayName().getString());
+        Voicechat.LOGGER.info("Sent secret to {}", player.getDisplayName().getString());
+    }
+
+    public void playerLoggedIn(ServerPlayer serverPlayer) {
+        if (!Voicechat.SERVER_CONFIG.forceVoiceChat.get()) {
+            return;
+        }
+
+        Timer timer = new Timer("%s-login-timer".formatted(serverPlayer.getGameProfile().getName()), true);
+        timer.schedule(new TimerTask() {
+            @Override
+            public void run() {
+                timer.cancel();
+                timer.purge();
+                if (!serverPlayer.server.isRunning()) {
+                    return;
+                }
+                if (!serverPlayer.connection.connection.isConnected()) {
+                    return;
+                }
+                if (!isCompatible(serverPlayer)) {
+                    serverPlayer.server.execute(() -> {
+                        serverPlayer.connection.disconnect(
+                                new TextComponent("You need %s %s to play on this server".formatted(
+                                        CommonCompatibilityManager.INSTANCE.getModName(),
+                                        CommonCompatibilityManager.INSTANCE.getModVersion()
+                                ))
+                        );
+                    });
+                }
+            }
+        }, Voicechat.SERVER_CONFIG.loginTimeout.get());
     }
 
     public void playerLoggedOut(ServerPlayer player) {